Plugins: OpenBCI, scale only EEG...
Former-commit-id: b757f32737999f50fae8218781b1e89dfb46757e

+++ contrib/plugins/server-drivers/openbci/src/ovasCDriverOpenBCI.h
@@ -102,13 +102,13 @@ namespace OpenViBEAcquisitionServer
 		OpenViBE::int16  m_i16LastPacketNumber;
 		OpenViBE::uint16 m_ui16Switches;
 
-		std::vector < std::vector < OpenViBE::int32 > > m_vChannelBuffer; // buffer to store channels & chunks
+		std::vector < std::vector < OpenViBE::float32 > > m_vChannelBuffer; // buffer to store channels & chunks
 		// buffer to store sample coming from OpenBCI -- filled by parseByteP2(), passed to handleCurrentSample()
-		std::vector < OpenViBE::int32 > m_vSampleEEGBuffer;
-		std::vector < OpenViBE::int32 > m_vSampleAccBuffer;
+		std::vector < OpenViBE::float32 > m_vSampleEEGBuffer;
+		std::vector < OpenViBE::float32 > m_vSampleAccBuffer;
 		// buffers to store & merge daisy samples
-		std::vector < OpenViBE::int32 > m_vSampleEEGBufferDaisy; 
-		std::vector < OpenViBE::int32 > m_vSampleAccBufferDaisy;
+		std::vector < OpenViBE::float32 > m_vSampleEEGBufferDaisy; 
+		std::vector < OpenViBE::float32 > m_vSampleAccBufferDaisy;
 
 		OpenViBE::CString m_sTTYName;
 	};
